#28: Moved source files
