+++ lang-portal/vocab-importer/src/data_processor.py
