+++ lang-portal/vocab-importer/src/sql_generator.py
