+++ lang-portal/vocab-importer/test/test_sql_generator.py
@@ -1,5 +1,5 @@
 import pytest
-from sql_generator import SQLGenerator
+from src.sql_generator import SQLGenerator
 
 def test_generate_kanji_insert():
     generator = SQLGenerator()
